Perform Contrast Test

This function performs a contrast test on a given dataset to compare the means of multiple variables using weighted contrasts. It calculates several statistics, including the t-value, F-value, p-value, confidence intervals, and effect sizes (Cohen's f and partial eta squared).

Usage

contrast_test(
  dat,
  weight = NULL,
  paired = FALSE,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  verbose = TRUE
)

Arguments

dat

A data frame or tibble containing the data to be analyzed. Each column represents a variable. For unbalanced data, please fill missing values with NA.

weight

A numeric vector of weights to apply to each variable in the contrast. The sum of the weights must be zero.

paired

A logical indicating whether the data is paired (TRUE) or unpaired (FALSE). Default is FALSE.

alternative

A character string specifying the type of test. Possible values are "two.sided" (default), "less", or "greater".

conf.level

A numeric value indicating the confidence level for the confidence intervals. Default is 0.95.

verbose

Logical. If TRUE, prints additional messages.

Value

A data frame containing the following columns:

weights

The weights applied to each variable in the contrast.

estimate

The weighted contrast estimate.

lower

The lower bound of the confidence interval.

upper

The upper bound of the confidence interval.

t

The t-value for the contrast.

F

The F-value for the contrast.

df_error

The degrees of freedom for the error.

SS_contrast

The sum of squares for the contrast component.

SS_error

The sum of squares for the error.

p

The p-value for the contrast test.

peta2

The partial eta squared effect size.

cohens_f

The Cohen's f effect size.

n_total

The total number of observations in the unpaired data.

n_total_na

The number of missing values in the unpaired data.

n_pair

The number of non-missing pairs in the paired data.

n_pair_na

The number of missing pairs in the paired data.

Details

This function performs a contrast analysis on a given dataset using the provided weights. It computes statistics such as the t-value, F-value, confidence intervals, and effect sizes (Cohen's f and partial eta squared). If the data is paired, pair-wise exclusion is applied. The function supports two-tailed and one-tailed tests depending on the value of the alternative parameter.

Examples

# Example of linear trend analysis with unpaired data
dat <- data.frame(A = rnorm(100), B = rnorm(100), C = rnorm(100))
weight <- c(1, -1, 0)
contrast_test(dat, weight)
#>   weights   estimate      lower     upper          t          F df_error
#> 1  1,-1,0 -0.0440676 -0.3332109 0.2450757 -0.2999354 0.08996123      297
#>   SS_contrast SS_error         p       peta2   cohens_f n_total n_total_na
#> 1  0.09709766 320.5604 0.7644363 0.000302808 0.01740402     300          0

# Example with paired data
dat <- data.frame(A = rnorm(50), B = rnorm(50), C = rnorm(50))
weight <- c(1, -2, 1)
contrast_test(dat, weight, paired = TRUE)
#>   weights   estimate     lower     upper         t        F df_error
#> 1  1,-2,1 -0.4605356 -1.135526 0.2144551 -1.353971 1.833238       98
#>   SS_contrast SS_error         p    peta2  cohens_f n_pair n_pair_na
#> 1    1.767442 94.48271 0.1788597 0.018363 0.1367717     50         0