Perform ANOVA using the anovakun procedure (version 4.8.9)

This function performs analysis of variance (ANOVA) based on the anovakun version 4.8.9, originally developed by Prof. Ryuta Iseki (Taisho University, Japan).

Usage

anovakun_(
  dataset,
  design,
  ...,
  long = FALSE,
  type2 = FALSE,
  nopost = FALSE,
  tech = FALSE,
  data.frame = FALSE,
  copy = FALSE,
  holm = FALSE,
  hc = FALSE,
  s2r = FALSE,
  s2d = FALSE,
  fs1 = FALSE,
  fs2r = FALSE,
  fs2d = FALSE,
  welch = FALSE,
  criteria = FALSE,
  lb = FALSE,
  gg = FALSE,
  hf = FALSE,
  cm = FALSE,
  auto = FALSE,
  mau = FALSE,
  har = FALSE,
  iga = FALSE,
  ciga = FALSE,
  eta = FALSE,
  peta = FALSE,
  geta = NA,
  eps = FALSE,
  peps = FALSE,
  geps = NA,
  omega = FALSE,
  omegana = FALSE,
  pomega = FALSE,
  gomega = NA,
  gomegana = NA,
  prep = FALSE,
  nesci = FALSE,
  besci = FALSE,
  cilmd = FALSE,
  cilm = FALSE,
  cind = FALSE,
  cin = FALSE,
  ciml = FALSE,
  cipaird = FALSE,
  cipair = FALSE,
  bgraph = c(NA, NA)
)

Arguments

dataset

A data frame containing the input data.

design

A character string specifying the experimental design (e.g., "As", "ABs", "sA", "sAB", "AsB", etc.).

...

Numbers of levels of each factor.

long

Logical. Whether the data is in long format.

type2

Logical. Whether to use Type II sums of squares (default is FALSE for Type III).

nopost

Logical. If TRUE, skips post hoc tests.

tech

Logical. If TRUE, returns raw lists of results instead of printing.

data.frame

Logical. If TRUE, returns the reformatted data frame.

copy

Logical. If TRUE, copies the output to the clipboard.

holm, ..., bgraph

Additional flags controlling specific analyses (see documentation).

Value

Depending on the arguments, either printed output or a list of ANOVA results.

Details

In this implementation, the code of anovakun version 4.8.9 has been included as-is with no modification to its computational logic or output. However, for the purpose of maintainability and modularity, internal helper functions have been separated into a different file.

Additionally, all original Japanese-language comments have been removed from the source code to maintain stylistic consistency with this package and to simplify documentation.

Redistribution and minor modifications of the original function are permitted under the terms indicated by the original author, provided that such modifications are clearly stated. This implementation complies with that policy.

For methodological details and rationale, please refer to the original documentation of anovakun version 4.8.9.

See also

Examples

 data_snakemr %>%
   anovakun_("sABC", 2, 2, 5, long = TRUE)
#> Error in anovakun_(., "sABC", 2, 2, 5, long = TRUE): could not find function "anovakun_"