Get a list of currently loaded packages and their versions

This function retrieves the names and versions of all currently loaded R packages. It returns a data frame with the package names in the first column and their respective versions in the second column.

Usage

loaded_packages_version()

Value

A data frame with two columns:

package

The name of the loaded package.

version

The version of the loaded package.

Examples

# Get the list of loaded packages with versions
loaded_packages_version()
#>         package    version
#> 1   BayesFactor 0.9.12.4.7
#> 2          MASS     7.3.64
#> 3        Matrix      1.7.2
#> 4  MatrixModels      0.5.4
#> 5            R6      2.6.1
#> 6          Rcpp     1.0.14
#> 7          base      4.4.3
#> 8    bayestestR     0.15.2
#> 9         bslib      0.9.0
#> 10       cachem      1.1.0
#> 11          cli      3.6.4
#> 12         coda   0.19.4.1
#> 13     compiler      4.4.3
#> 14         curl      6.2.2
#> 15     datasets      4.4.3
#> 16   datawizard      1.0.2
#> 17         desc      1.4.3
#> 18     devtools      2.4.5
#> 19       digest     0.6.37
#> 20      downlit      0.4.4
#> 21        dplyr      1.1.4
#> 22   effectsize      1.0.0
#> 23     ellipsis      0.3.2
#> 24     evaluate      1.0.3
#> 25        fansi      1.0.6
#> 26      fastmap      1.2.0
#> 27  fontawesome      0.5.3
#> 28           fs      1.6.5
#> 29     generics      0.1.3
#> 30         glue      1.8.0
#> 31    grDevices      4.4.3
#> 32     graphics      4.4.3
#> 33         grid      4.4.3
#> 34    htmltools    0.5.8.1
#> 35  htmlwidgets      1.6.4
#> 36       httpuv     1.6.15
#> 37        httr2      1.1.2
#> 38      insight      1.1.0
#> 39    jquerylib      0.1.4
#> 40     jsonlite      2.0.0
#> 41        knitr       1.50
#> 42        later      1.4.1
#> 43      lattice     0.22.6
#> 44    lifecycle      1.0.4
#> 45     magrittr      2.0.3
#> 46      memoise      2.0.1
#> 47      methods      4.4.3
#> 48         mime       0.13
#> 49       miniUI    0.1.1.1
#> 50     mutolabr      1.3.0
#> 51      mvtnorm      1.3.3
#> 52     parallel      4.4.3
#> 53   parameters     0.24.2
#> 54      pbapply      1.7.2
#> 55       pillar     1.10.2
#> 56     pkgbuild      1.4.7
#> 57    pkgconfig      2.0.3
#> 58      pkgdown      2.1.1
#> 59      pkgload      1.4.0
#> 60        ppcor        1.1
#> 61      profvis      0.4.0
#> 62     promises      1.3.2
#> 63        purrr      1.0.4
#> 64     rappdirs      0.3.3
#> 65      remotes      2.5.0
#> 66        rlang      1.1.5
#> 67    rmarkdown       2.29
#> 68   rstudioapi     0.17.1
#> 69         sass      0.4.9
#> 70  sessioninfo      1.2.3
#> 71        shiny     1.10.0
#> 72        stats      4.4.3
#> 73      stringi      1.8.7
#> 74      stringr      1.5.1
#> 75       tibble      3.2.1
#> 76        tidyr      1.3.1
#> 77   tidyselect      1.2.1
#> 78        tools      4.4.3
#> 79   urlchecker      1.0.1
#> 80      usethis      3.1.0
#> 81        utils      4.4.3
#> 82        vctrs      0.6.5
#> 83      whisker      0.4.1
#> 84        withr      3.0.2
#> 85         xfun       0.52
#> 86         xml2      1.3.8
#> 87       xtable      1.8.4
#> 88         yaml     2.3.10